bench(moq-json): use Criterion instead of a hand-rolled timing harness
Replace the home-grown Instant-based loop in examples/bench.rs with a proper
Criterion benchmark under benches/codec.rs (cargo bench -p moq-json). It drives
the real public Producer/Consumer, so it benchmarks the actual serializer-backed
merge-patch path, and compares against a DEFLATE-only baseline plus a DEFLATE
level sweep, with throughput reported per workload.

The wire-size demonstration already lives in examples/telemetry.rs, so the
deleted example loses nothing there.
